- remove cyrillic description - html formatting - change search phrases logic

+++ playwright-scrapper/src/main/java/com/playwright/scrapper/model/report/PageReport.java
@@ -5,5 +5,5 @@ import java.util.Map;
 
 public record PageReport(
         PerformanceInfo performanceInfo,
-        Map<String, Integer> searchPhrases) {
+        Map<String, Long> searchPhrases) {
 }
